feat: 规划P1问题修复顺序
